Fix MySQL 5.7 migration error
The migration in `add_time_zone_awareness.py` fails on MySQL 5.7
due to primary key columns not having `nullable` explicitly set to
False. This value defaults to true, and hence the migration
will fail with a MySQL error.

This explicitly sets `nullable` to False for the
alter column statements of `sla_miss.execution_date`
and `task_instance.execution_date`.